// ui/src/generated/api/model/dtoSetAccessResponse.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Vertex OJ API
 * Public application endpoints and the authenticated Judge worker protocol.
 * OpenAPI spec version: 1.0
 */
import type { DtoSetAccessResponseRole } from "./dtoSetAccessResponseRole.ts";

export interface DtoSetAccessResponse {
  groupId?: string;
  groupName?: string;
  id: number;
  role: DtoSetAccessResponseRole;
  userId?: string;
  username?: string;
}